Add post on hosting Overleaf Locally
The guide covers generating configs, starting the server, creating an
account and a project, as well as updating the TeXLive distribution so
that all the packages the user wants are installed inside the container.
It also shows how to persist this container as a new image and always
have overleaf toolkit boot this image by default.